+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- Mask: add new prop `willChangeTransform` default true which can turn off `willChange:transform` property in CSS (#494)
+
 ### Patch
 
 </details>